Adjust fprintf arguments type matching format specifier in test_stack
* tests/test_stack.c (main): Cast "code" argument of fprintf calls to
unsigned int to match printf format specifier.

+++ tests/test_stack.c
@@ -260,7 +260,7 @@ int main(int argc, char **argv)
                                   (void *)(size_t)i);
 #         endif
           if (code != 0) {
-            fprintf(stderr, "Thread creation failed %u\n", code);
+            fprintf(stderr, "Thread creation failed %u\n", (unsigned)code);
             exit(3);
           }
         }
@@ -277,7 +277,7 @@ int main(int argc, char **argv)
             code = pthread_join(thread[i], 0);
 #         endif
           if (code != 0) {
-            fprintf(stderr, "Thread join failed %u\n", code);
+            fprintf(stderr, "Thread join failed %u\n", (unsigned)code);
             abort();
           }
         }
